\name{pesco-package}
\alias{pesco-package}
\alias{pesco}
\docType{package}
\title{
Data fusion for air quality data
}
\description{
PESCO stands for "Post-processing and Evaluation
             with Statistical methods of a Chemistry-transport-model
             Output". The package provides functions to perform data
             fusion for air quality data, correcting the output of a
             deterministic CTM with observed data, with a Trans-Gaussian
             Kriging approach.
}
\details{
\tabular{ll}{
Package: \tab pesco\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2015-03-26\cr
License: \tab GPL-2\cr
}
%~~ An overview of how to use the package, including the most ~~
%~~ important functions ~~
}
\author{
Lucia Paci, Giovanni Bonafe'

Maintainer: G.Bonafe' <gbonafe@arpa.emr.it>
}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file ~~
%~~ KEYWORDS in the R documentation directory ~~
%\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
